src/decoder.py
import numpy as np
import soundfile as sf

def decode_audio_to_image(audio_file, sample_rate=44100):
    """
    Stub function to decode an SSTV audio file back into an image.
    This function currently returns a dummy black image.
    """
    # Read audio file (actual decoding logic to be implemented)
    audio_signal, sr = sf.read(audio_file)
    # For demonstration, create a dummy image (black image)
    height = 256
    width = 320
    image_array = np.zeros((height, width), dtype=np.uint8)
    return image_array
