requirements.txt
setuptools
numpy
opencv-python
pyaudio
sounddevice
scipy
matplotlib
Pillow
wave